Add files via upload
This is a baby example of Random Scan Gibbs sampling. Suppose that you can sample from a univariate normal distribution, the script shows how you can sample from a bivariate normal distribution using Gibbs sampling through a random scan of the kernels, which are the full conditionals. Try it out with \rho=0.99 and a big number of iterations.